--- a/examples/reduce.mojo
+++ b/examples/reduce.mojo
@@ -21,9 +21,6 @@ from algorithm import sum
 from random import rand
 from memory.buffer import Buffer
 
-# Change these numbers to reduce on different sizes
-alias numElemSmall = 1 << 21
-alias numElemLarge = 1 << 29
 
 # Simple array struct
 struct ArrayInput:
@@ -69,95 +66,49 @@ fn benchmark_naive_reduce_sum(size: Int) -> Float32:
     return mySum
 
 
-fn benchmark_stdlib_reduce_sum(size: String) -> Float32:
+fn benchmark_stdlib_reduce_sum(size: Int) -> Float32:
     # Allocate a Buffer and then use the Mojo stdlib Reduction class
-    if size == 'small':
+    # TODO: Use globals
+    # alias numElem = size
+    alias numElem = 1 << 29
+    # Can use either stack allocation or heap
+    # see stackalloc
+    # var A = Buffer[numElem, DType.float32].stack_allocation()
+    # see heapalloc
+    var B = DTypePointer[DType.float32].alloc(numElem)
+    var A = Buffer[numElem, DType.float32](B)
 
-        alias num_elem_small = numElemSmall
-        # Can use either stack allocation or heap
-        # see stackalloc
-        # var A = Buffer[num_elem_small, DType.float32].stack_allocation()
-        # see heapalloc
-        var B = DTypePointer[DType.float32].alloc(num_elem_small)
-        var A = Buffer[num_elem_small, DType.float32](B)
+    # initialize buffer
+    for i in range(numElem):
+        A[i] = Float32(i)
 
-        # initialize buffer
-        for i in range(num_elem_small):
-            A[i] = Float32(i)
+    # Prevent DCE
+    var mySum: Float32 = 0.0
+    print("Computing reduction sum for array num elements: ", size)
 
-        # Prevent DCE
-        var mySum: Float32 = 0.0
-        print("Computing reduction sum for array num elements: ", num_elem_small)
+    @always_inline
+    @parameter
+    fn test_fn():
+        mySum = sum[numElem, DType.float32](A)
 
-        @always_inline
-        @parameter
-        fn test_fn():
-            mySum = sum[num_elem_small, DType.float32](A)
-
-        let bench_time = Float64(Benchmark().run[test_fn]())
-        return mySum
-
-    else:
-        alias num_elem_large = numElemLarge
-        # Can use either stack allocation or heap
-        # see stackalloc
-        # var A = Buffer[num_elem_large, DType.float32].stack_allocation()
-        # see heapalloc
-        var B = DTypePointer[DType.float32].alloc(num_elem_large)
-        var A = Buffer[num_elem_large, DType.float32](B)
-
-        # initialize buffer
-        for i in range(num_elem_large):
-            A[i] = Float32(i)
-
-        # Prevent DCE
-        var mySum: Float32 = 0.0
-        print("Computing reduction sum for array num elements: ", num_elem_large)
-
-        @always_inline
-        @parameter
-        fn test_fn_large():
-            mySum = sum[num_elem_large, DType.float32](A)
-
-        let bench_time = Float64(Benchmark().run[test_fn_large]())
-        return mySum
+    let bench_time = Float64(Benchmark().run[test_fn]())
+    return mySum
 
 
 fn main():
+    # Number of array elements
+    let size = 1 << 29
     print("# Reduction sum across a large array. The naive algorithm's ")
     print("# computation time scales with the size of the array; while Mojo ")
     print("# exhibits significantly better scaling...")
-
-    # Number of array elements
-    let size_small: Int = numElemSmall
-    let size_large: Int = numElemLarge
-
-    # smaller scale
-
-    print("At a smaller scale", size_small, "elements.")
     var eval_begin: Float64 = now()
-    var sum = benchmark_naive_reduce_sum(size_small)
+    var sum = benchmark_naive_reduce_sum(size)
     var eval_end: Float64 = now()
     var execution_time = Float64((eval_end - eval_begin)) / 1e6
     print("Completed naive reduction sum: ", sum, " in ", execution_time, "ms")
 
     eval_begin = now()
-    sum = benchmark_stdlib_reduce_sum('small')
-    eval_end = now()
-    execution_time = Float64((eval_end - eval_begin)) / 1e6
-    print("Completed stdlib reduction sum: ", sum, " in ", execution_time, "ms")
-
-    # larger scale
-
-    print("At a larger scale", size_large, "elements.")
-    eval_begin= now()
-    sum = benchmark_naive_reduce_sum(size_large)
-    eval_end= now()
-    execution_time = Float64((eval_end - eval_begin)) / 1e6
-    print("Completed naive reduction sum: ", sum, " in ", execution_time, "ms")
-
-    eval_begin = now()
-    sum = benchmark_stdlib_reduce_sum('large')
+    sum = benchmark_stdlib_reduce_sum(size)
     eval_end = now()
     execution_time = Float64((eval_end - eval_begin)) / 1e6
     print("Completed stdlib reduction sum: ", sum, " in ", execution_time, "ms")