* original `naive` and `stblib` reduce is not fair: `non-stblib` reduce computed on `2^21`, whilst `stblib` reduce computed on `2^30`. This makes this fair.
* `naive` and `stblib` reduce both computed on size of `2^29` elements, around 18X speedup on my laptop.

Before:
```
# Reduction sum across a large array. The naive algorithm's 
# computation time scales with the size of the array; while Mojo 
# exhibits significantly better scaling...
Computing reduction sum for array num elements:  2097152
Completed naive reduction sum:  0.0  in  1273.46117 ms                       # <---------- notice this
Computing reduction sum for array num elements:  2097152
Completed stdlib reduction sum:  5.7667653146037453e+17  in  1944.556012 ms                      # <---------- notice this
```
After
```
# Reduction sum across a large array. The naive algorithm's 
# computation time scales with the size of the array; while Mojo 
# exhibits significantly better scaling...
Computing reduction sum for array num elements:  536870912
Completed naive reduction sum:  0.0  in  27702.951013000002 ms
Computing reduction sum for array num elements:  536870912
Completed stdlib reduction sum:  1.4433103595228365e+17  in  1568.946565 ms
```